google_api_tools/requirements.txt
# Google API dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Date/time handling
pytz>=2023.3
python-dateutil>=2.8.2

# Database
SQLAlchemy>=2.0.0

# AI/ML (for YouTube responder)
openai>=1.0.0

# Utilities
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Testing
pytest>=7.3.1
